Avoid flickering and a race condition on tasks overview page
Flickering was caused by multiple practice overview requests and by
allowing to show the old state of the table before the new state was
fetched. This commit also implement a simple "loading page" (with a
loading indicator on a dark background).

Race condition was allowed for fetching static data and practice
overview: if the practice overview arrived sooner, it was discarded,
because it couldn't be merged into tasks. This was solved by enforcing
order of loading (first static data, only then practice overview).